SDP-138 Testing and refactoring finished
