+++ src/entity/FinalBoss.java
@@ -1,7 +1,6 @@
 package entity;
 
 import java.awt.Color;
-import java.util.HashSet;
 import java.util.Set;
 
 import audio.SoundManager;
@@ -12,187 +11,102 @@ import engine.GameState;
 
 public class FinalBoss extends Entity implements BossEntity {
 
+    /** Current health points of the final boss */
     private int healPoint;
+    /** Maximum health points of the final boss */
     private int maxHp;
+    /** Score value awarded when the final boss is destroyed */
     private final int pointValue;
+    /** Flag indicating whether the final boss has been destroyed */
     private boolean isDestroyed;
     /** Distance between normal state position and Power-up state position*/
-    private final static int OFFSET = 30;
+    public final static int OFFSET = 30;
     /** Normal width */
-    private final static int NORMAL_WIDTH = 100;
+    public final static int NORMAL_WIDTH = 100;
     /** Normal height */
-    private final static int NORMAL_HEIGHT = 80;
+    public final static int NORMAL_HEIGHT = 80;
     /** power-up width */
-    private final static int POWERUP_WIDTH = 160;
+    public final static int POWERUP_WIDTH = 160;
     /** power-up height */
-    private final static int POWERUP_HEIGHT = 140;
+    public final static int POWERUP_HEIGHT = 140;
     /** Constant about phase 2 HP threshold */
 	public static final double PHASE_2_HP_THRESHOLD = 0.7;
 	/** Constant about phase 3 HP threshold */
 	public static final double PHASE_3_HP_THRESHOLD = 0.3;
-    /** Horizontal speed for zigzag moving */
-    private int zigzagSpeed;
-    /** Speed for dash pattern */
-    private int dashSpeed;
-    /** Cool time for dash pattern */
-    private int dashCooldownTime;
-    /** Cool time for dash charging */
-    private int dashChargeCooldowntime;
-    /** Cool time for shoot1 pattern */
-    private int shoot1Cooldowntime;
-    /** Cool time for laser pattern */
-    private int laserCooldowntime;
-    /** Cool time for laser duration */
-    private int laserDurationtime;
+    /** Manager of controling Attack logic */
+    private BossAttackManager attackManager;
+    /** Manager of controling Movement logic */
+    private BossMovementManager movementManager;
     /** The State of Dash Power Up */
     private boolean isPowerUp = false;
     /** The difficulty level of boss */
     private int difficulty;
-    /** The Ship of player */
-    private Ship player;
-    /** The Fixed Position X of player ship when start dash pattern*/
-    private int fixedPositionX;
-    /** The Fixed Position Y of player ship when start dash pattern*/
-    private int fixedPositionY;
-    /** for move pattern */
-    private int zigDirection = 1;
-    /** for move pattern */
-    private boolean goingDown = true;
-    /** Dash Velocity X */
-    private double dashVelX = 0;
-    /** Dash Velocity Y */
-    private double dashVelY = 0;
-    /** The Decimal Point of dashVelX */
-    private double subDashVelX;
-    /** The Decimal Point of dashVelY */
-    private double subDashVelY;
-    /** Degree of shaking X axis */
-    private int shakeOffX = 0;
-    /** Degree of shaking Y axis */
-    private int shakeOffY = 0;
     /** Animation cool down */
     private Cooldown animationCooldown;
     /** PowerUp animation cool down */
     private Cooldown animationPowerUpCooldown;
-    /** Shoot1's cool down */
-    private Cooldown shootCooldown1;
-    /** Shoot2's cool down */
-    private Cooldown shootCooldown2;
-    /** Shoot3's cool down */
-    private Cooldown shootCooldown3;
-    /** Shoot4's cool down */
-    private Cooldown shootCooldown4;
-    /** Cooldown for charging of dash */
-    private Cooldown dashChargeCooldown;
-    /** Cooldown for charging of laser */
-    private Cooldown chargeLaserCooldown;
-    /** Cooldown for laser pattern */
-    private Cooldown laserCooldown;
-    /** Dashing's cool down */
-    private Cooldown dashCooldown;
-    /** Duration of Laser cool down */
-    private Cooldown laserDuration;
-
-    /** Dash Pattern State */
-    private enum DashPattern {
-        NONE,
-        FOLLOWING,
-        CHARGING,
-        DASHING,
-    }
-
-    /** Laser Pattern State */
-    private enum LaserPattern {
-        NONE,
-        IDLE,
-        CHARGING,
-        FIRING,
-    }
-
-    /** Dash Pattern Current State */
-    private DashPattern dashOption;
-    /** Laser Pattern Current State */
-    private LaserPattern laserOption;
-
+    /** Screen width */
     private int screenWidth;
+    /** Screen height */
     private int screenHeight;
+    
     /** basic attribute of final boss */
-
     public FinalBoss(int positionX, int positionY, int screenWidth, int screenHeight, Ship ship, int difficulty){
 
         super(positionX,positionY,100,80, Color.RED);
+        
         this.difficulty = difficulty;
+
         if (GameState.isDecreaseEnemyPower()) {
             this.healPoint = 1;
         } else {
             switch(this.difficulty) {
                 case 1: // easy
                     this.healPoint = 60;
-                    this.zigzagSpeed = 3;
-                    this.dashSpeed = 8;
-                    this.dashCooldownTime = 3000;
-                    this.dashChargeCooldowntime = 1200;
-                    this.shoot1Cooldowntime = 2000;
-                    this.laserCooldowntime = 3000;
-                    this.laserDurationtime = 1600;
                     break;
                 case 2: // normal
                     this.healPoint = 90;
-                    this.zigzagSpeed = 4;
-                    this.dashSpeed = 10;
-                    this.dashCooldownTime = 3000;
-                    this.dashChargeCooldowntime = 1000;
-                    this.shoot1Cooldowntime = 1400;
-                    this.laserCooldowntime = 3000;
-                    this.laserDurationtime = 1600;
                     break;
                 case 3: // hard
                     this.healPoint = 120;
-                    this.zigzagSpeed = 5;
-                    this.dashSpeed = 14;
-                    this.dashCooldownTime = 2000;
-                    this.dashChargeCooldowntime = 500;
-                    this.shoot1Cooldowntime = 900;
-                    this.laserCooldowntime = 3000;
-                    this.laserDurationtime = 1600;
                     break;
                 default:
                     break;
             }
         }
+
         this.maxHp = healPoint;
         this.pointValue = 1000;
         this.spriteType = DrawManager.SpriteType.FinalBoss1;
         this.isDestroyed = false;
         this.screenWidth = screenWidth;
         this.screenHeight = screenHeight;
-        this.player = ship;
-        this.dashOption = DashPattern.NONE;
-        this.laserOption = LaserPattern.NONE;
 
         this.animationCooldown = Core.getCooldown(500);
         this.animationPowerUpCooldown = Core.getCooldown(250);
-        this.shootCooldown1 = Core.getCooldown(this.shoot1Cooldowntime);
-        this.shootCooldown2 = Core.getCooldown(400);
-        this.shootCooldown3 = Core.getCooldown(300);
-        this.shootCooldown4 = Core.getCooldown(5000);
-        this.chargeLaserCooldown = Core.getCooldown(1500);
-        this.dashChargeCooldown = Core.getCooldown(this.dashChargeCooldowntime);
-        this.dashCooldown = Core.getCooldown(this.dashCooldownTime);
-        this.laserCooldown = Core.getCooldown(this.laserCooldowntime);
-        this.laserDuration = Core.getCooldown(this.laserDurationtime);
-    }
 
-    /** for vibrant moving with final boss
-     * final boss spritetype is the same with special enemy and enemyshipA, because final boss spritetype have not yet implemented
-     * becasue final boss is single object, moving and shooting pattern are included in update methods
-     */
+        this.attackManager = new BossAttackManager(this);
+        this.movementManager = new BossMovementManager(this, ship);
+    }
+ 
     @Override
     public void update(){
         runAnimation();
-        movePattern();
+
+        this.movementManager.updateMovement();
     }
 
+    /** call attack logic function of final boss of attackManager */
+    public Set<BossAttack> processAttacks() {
+        return this.attackManager.processAttacks();
+    }
+
+    /** call checking function whether need to clear attacks */
+    public boolean shouldClearAttacks() {
+        return this.attackManager.shouldClearAttacks();
+    }
+
+    /** run animation of final boss */
     public void runAnimation() {
         if (this.isPowerUp) {
             if(this.animationPowerUpCooldown.checkFinished()){
@@ -233,6 +147,13 @@ public class FinalBoss extends Entity implements BossEntity {
         }
     }
 
+    /** move final boss by distanceX and distanceY */
+    @Override
+    public void move(int distanceX, int distanceY){
+        this.positionX += distanceX;
+        this.positionY += distanceY;
+    }
+
     /** decrease boss' healpoint */
     @Override
     public void takeDamage(int damage){
@@ -254,6 +175,7 @@ public class FinalBoss extends Entity implements BossEntity {
         return this.healPoint;
     }
 
+    /** get max hp of final boss */
     public int getMaxHp(){
         return  this.maxHp;
     }
@@ -263,212 +185,7 @@ public class FinalBoss extends Entity implements BossEntity {
         return this.pointValue;
     }
 
-    /** move simple */
-    @Override
-    public void move(int distanceX, int distanceY){
-        this.positionX += distanceX;
-        this.positionY += distanceY;
-    }
-    
-    /** movement pattern of final boss */
-    public void movePattern(){
-        if(this.healPoint > this.maxHp*PHASE_2_HP_THRESHOLD) {
-            if (this.difficulty == 1) {
-                this.move(0,0);
-            } else {
-                this.moveZigzag(2,1);
-            }
-        } else if (this.healPoint > this.maxHp*PHASE_3_HP_THRESHOLD) {
-            this.activateLaserPattern();
-
-            this.moveZigzag(this.zigzagSpeed, 3);
-        }
-        else {
-            this.deactivateLaserPattern();
-            this.activateDashPattern();
-
-            dashPattern(this.dashSpeed);
-        }
-    }
-
-    public Set<BossAttack> shootBossAttack() {
-        Set<BossAttack> attacks = new HashSet<>();
-
-        if(this.healPoint > this.maxHp*PHASE_2_HP_THRESHOLD) {
-            if (this.difficulty == 1) {
-                attacks.addAll(this.shoot1());
-                attacks.addAll(this.shoot2());
-            } else {
-                attacks.addAll(this.shoot3());
-            }
-        }else if (this.healPoint > this.maxHp*PHASE_3_HP_THRESHOLD) {
-            if (this.difficulty != 1 && !is_cleared) {
-                attacks.clear();
-                is_cleared = true;
-            } else {
-                attacks.addAll(this.shoot1());
-                attacks.addAll(this.shoot2());
-                attacks.addAll(this.laserShoot());
-        }
-        else {
-            if (this.difficulty != 1) {
-                attacks.addAll(this.shoot4());
-            }
-            attacks.addAll(this.shoot2());
-        }
-
-        return attacks;
-    }
-
-    /** preprocess before laser Pattern */
-    public void activateLaserPattern() {
-        if (this.laserOption == LaserPattern.NONE) {
-                this.laserOption = LaserPattern.IDLE;
-                this.laserCooldown.reset();
-        }
-    }
-
-    /** Set after finishing the laser pattern */
-    public void deactivateLaserPattern() {
-        if (this.laserOption != LaserPattern.NONE && this.laserDuration.checkFinished()) {
-                this.laserOption = LaserPattern.NONE;
-                this.setClear();
-        }
-    }
-
-    /** preprocess before dash Pattern */
-    public void activateDashPattern() {
-        if (this.dashOption == DashPattern.NONE && this.laserOption == LaserPattern.NONE) {
-                this.dashOption = DashPattern.FOLLOWING;
-                this.dashCooldown.reset();
-        }
-    }
-
-    /** Set after finishing the dash pattern */
-    public void deactivateDashPattern() {
-        if (this.dashOption != DashPattern.NONE) {
-                this.dashOption = DashPattern.NONE;
-                this.setClear();
-        }
-    }
-
-    /** move zigzag */
-    public void moveZigzag(int zigSpeed, int vertSpeed){
-        this.positionX += (this.zigDirection * zigSpeed);
-        if(this.positionX <= 0) {
-            this.positionX = 0;
-            this.zigDirection = 1;
-        } else if (this.positionX >= this.screenWidth - this.width) {
-            this.positionX = this.screenWidth - this.width;
-            this.zigDirection = -1;
-        }
-
-        if(goingDown) {
-            this.positionY += vertSpeed;
-            if (this.positionY >= screenHeight/2 - this.height) goingDown = false;
-        }
-        else {
-            this.positionY -= vertSpeed;
-            if(this.positionY <= 0) goingDown = true;
-        }
-    }
-
-    /** Manage the Dash Pattern */
-    public void dashPattern(int dashSpeed) {
-        switch (this.dashOption) {
-            case FOLLOWING:
-                this.follow(dashSpeed);
-                this.dashChargeCooldown.reset();
-                break;
-            case CHARGING:
-                this.charge();
-                break;
-            case DASHING:
-                dash();
-                break;
-            default:
-                break;
-        }
-    }
-
-    /** Fix the Position of Player Ship */
-    public void follow(int dashSpeed) {
-        this.dashCooldown.reset();
-        this.changeToPowerUpMode();
-
-        this.fixedPositionX = this.player.getPositionX() - this.getWidth()/2 + this.player.getWidth()/2; 
-        this.fixedPositionY = this.player.getPositionY() - this.getHeight()/2 + this.player.getHeight()/2;
-
-        double angle = Math.atan2(fixedPositionY - positionY, fixedPositionX - positionX);
-
-        this.dashVelX = dashSpeed * Math.cos(angle);
-        this.dashVelY = dashSpeed * Math.sin(angle);
-
-        this.dashOption = DashPattern.CHARGING;
-
-        this.subDashVelX = 0;
-        this.subDashVelY = 0;
-    }
-
-    /** Charge for a While Before Starting Dash Pattern */
-    public void charge() {
-        if (this.dashChargeCooldown.checkFinished()) {
-            this.dashOption = DashPattern.DASHING;
-
-            this.move(-shakeOffX, -shakeOffY);
-            this.shakeOffX = 0;
-            this.shakeOffY = 0;
-
-        } else {
-            this.move(-shakeOffX, -shakeOffY);
-
-            int shakeAmount = 2;
-            this.shakeOffX = java.util.concurrent.ThreadLocalRandom.current().nextInt(-1, 2) * shakeAmount;
-            this.shakeOffY = java.util.concurrent.ThreadLocalRandom.current().nextInt(-1, 2) * shakeAmount;
-
-            this.move(this.shakeOffX, this.shakeOffY);
-        }
-    }
-
-    /** Dash Pattern */
-    public void dash() {
-        double dx = fixedPositionX - positionX;
-        double dy = fixedPositionY - positionY;
-        double distanceSq = dx * dx + dy * dy;
-        double speedSq = dashVelX * dashVelX + dashVelY * dashVelY;
-        
-		if (distanceSq <= speedSq || !this.isPowerUp) {
-
-            if (this.isPowerUp) {
-                this.positionX = this.fixedPositionX;
-                this.positionY = this.fixedPositionY;
-                this.changeToPowerUpMode();
-            }
-
-            if (this.dashCooldown.checkFinished()) {
-                this.dashCooldown.reset();
-
-                this.dashOption = DashPattern.FOLLOWING;
-                this.dashVelX = 0;
-                this.dashVelY = 0;
-            }
-        } else {
-
-            this.subDashVelX += this.dashVelX;
-            this.subDashVelY += this.dashVelY;
-
-            int moveX = (int) this.subDashVelX;
-            int moveY = (int) this.subDashVelY;
-
-            if (moveX != 0 || moveY !=0) {
-                this.move(moveX, moveY);
-
-                this.subDashVelX -= moveX;
-                this.subDashVelY -= moveY;
-            }
-        }
-    }
-
+    /** change final boss state between power-up and normal*/
     public void changeToPowerUpMode() {
         if (this.isPowerUp) {
             this.isPowerUp = false;
@@ -491,125 +208,36 @@ public class FinalBoss extends Entity implements BossEntity {
         }
     }
 
-    /** first shooting pattern of final boss */
-    public Set<BossBullet> shoot1(){
-        if(this.shootCooldown1.checkFinished() && (this.laserOption == LaserPattern.IDLE || this.laserOption == LaserPattern.NONE)) {
-            this.shootCooldown1.reset();
-            Set<BossBullet> bullets = new HashSet<>();
-            int arr[] = {0,1,-1,2,-2};
-            for (int i : arr){
-                BossBullet bullet = new BossBullet(this.getPositionX() + this.getWidth() / 2 - 3,this.getPositionY() + this.getHeight(), i,4,6,10,Color.yellow);
-                bullets.add(bullet);
-            }
-            return bullets;
-        }
-        return java.util.Collections.emptySet();
-    }
-
-    /** second shooting pattern of final boss */
-    public Set<BossBullet> shoot2() {
-        if (this.shootCooldown2.checkFinished()) {
-            this.shootCooldown2.reset();
-            Set<BossBullet> bullets = new HashSet<>();
-            int randomX = (int) (Math.random() * screenWidth);
-            BossBullet bullet = new BossBullet(randomX, 1, 0, 2,6,10,Color.yellow);
-            bullets.add(bullet);
-            return bullets;
-        }
-        return java.util.Collections.emptySet();
-    }
-    
-    /** third shooting pattern of final boss */
-    public Set<BossBullet> shoot3() {
-        Set<BossBullet> bullets = new HashSet<>();
-        if (this.shootCooldown3.checkFinished()) {
-            this.shootCooldown3.reset();
-         // if (!(this.getPositionX() == 0 || this.getPositionX() == 400)){
-                BossBullet bullet1 = new BossBullet(this.getPositionX() + this.getWidth() / 2 - 3 + 70, this.positionY, 0, 5,6,10,Color.blue);
-                BossBullet bullet2 = new BossBullet(this.getPositionX() + this.getWidth() / 2 - 3 - 70, this.positionY, 0, 5,6,10,Color.blue);
-                bullets.add(bullet1);
-                bullets.add(bullet2);
-         // }
-        }
-        return bullets;
-    }
-
-    public Set<BossBullet> shoot4(){
-        if(this.shootCooldown4.checkFinished() && this.laserDuration.checkFinished()){
-            this.shootCooldown4.reset();
-            Set<BossBullet> bullets = new HashSet<>();
-            int bulletCount = 12;
-            int speed;
-            if (this.difficulty == 3) {
-                speed = 7;
-            } else {
-                speed = 5;
-            }
-
-            for (int i=0; i<bulletCount; i++) {
-                double angle = 2* Math.PI * i / bulletCount;
-
-                double velX = Math.cos(angle) * speed;
-                double velY = Math.sin(angle) * speed;
-
-                BossBullet bullet = new BossBullet(this.getPositionX() + this.getWidth() / 2 - 3, this.getPositionY() + this.getHeight() / 2,
-                                    (int)velX, (int)velY, 6, 10, Color.BLUE);
-                bullets.add(bullet);
-            }
-            return bullets;
-        }
-        return java.util.Collections.emptySet();
-    }
-
-    /** Laser firing pattern of final boss */
-    public Set<BossLaser> laserShoot() {
-        Set<BossLaser> lasers = new HashSet<>();
-
-        switch (this.laserOption) {
-            case IDLE:
-                if (this.laserCooldown.checkFinished()) {
-                    this.laserOption = LaserPattern.CHARGING;
-                    this.chargeLaserCooldown.reset();
-
-                    this.changeToPowerUpMode();
-                }
-                break;
-
-            case CHARGING:
-                if (this.chargeLaserCooldown.checkFinished()) {
-                    this.laserOption = LaserPattern.FIRING;
-                    this.laserDuration.reset();
-
-                    for (int i=0; i<9; i++) {
-                        BossLaser laser = new BossLaser(this.getPositionX()+OFFSET, this.getPositionY()+POWERUP_HEIGHT + NORMAL_HEIGHT*i, NORMAL_WIDTH, NORMAL_HEIGHT, this, i, this.laserDuration);
-                        lasers.add(laser);
-                    }
-                }
-                break;
-
-            case FIRING:
-                if (this.laserDuration.checkFinished()) {
-                    this.laserOption = LaserPattern.IDLE;
-                    this.laserCooldown.reset();
-
-                    this.changeToPowerUpMode();
-                }
-                break;
-
-            default:
-                break;
-            }
-        return lasers;
-    }
-
+    /** set clear state after changing to boss phase */
     public void setClear() {
         if (this.isPowerUp) {
             this.changeToPowerUpMode();
         }
     }
 
+    /** check whether laser is activated */
+    public boolean isLaserActivate() {
+        return this.attackManager.isLaserActivate();
+    }
+
+    /** get screen width */
+    public int getScreenWidth() {
+        return this.screenWidth;
+    }
+
+    /** get screen height */
+    public int getScreenHeight() {
+        return this.screenHeight;
+    }
+
+    /** get difficulty level */
     public int getDifficulty() {
         return this.difficulty;
+    }
+
+    /** get the state of Power Up */
+    public boolean getIsPowerUp() {
+        return this.isPowerUp;
     }
 
     /** flag final boss' destroy */
