+++ src/screen/GameScreen.java
@@ -68,9 +68,6 @@ public class GameScreen extends Screen {
 
     private static final int ITEMS_SEPARATION_LINE_HEIGHT = 600;
 
-    private static final int TAKE_LASER_DAMAGE_TIME = 3000;
-
-
     /**
      * Current level data (direct from Level system).
      */
@@ -100,13 +97,8 @@ public class GameScreen extends Screen {
      */
     private Cooldown bossExplosionCooldown;
     /**
-     * Time until the player can take damage again.
-     */
-    private Cooldown takeLaserDamageCooldown;
-    /**
      * Time from finishing the level to screen change.
      */
-
     private EnemyShipChaserFormation chaserFormation;
 
     private Cooldown screenFinishedCooldown;
@@ -171,10 +163,6 @@ public class GameScreen extends Screen {
      * bossBullets carry bullets or lasers which Boss fires
      */
     private Set<BossAttack> bossAttacks;
-    /**
-     * bossLasers carry lasers which Boss fires
-     */
-    private boolean is_cleared = false;
     /**
      * Timer to track elapsed time.
      */
@@ -279,8 +267,6 @@ public class GameScreen extends Screen {
         enemyShipSpecialFormation.attach(this);
         this.bossExplosionCooldown = Core
                 .getCooldown(BOSS_EXPLOSION);
-        this.takeLaserDamageCooldown = Core
-                .getCooldown(TAKE_LASER_DAMAGE_TIME);
         this.screenFinishedCooldown = Core.getCooldown(SCREEN_CHANGE_INTERVAL);
         this.bullets = new HashSet<Bullet>();
         this.dropItems = new HashSet<DropItem>();
@@ -948,7 +934,6 @@ public class GameScreen extends Screen {
                 this.logger.warning("Unknown bossId: " + bossName);
                 break;
         }
-        this.is_cleared = false;
     }
 
     /**
@@ -962,73 +947,14 @@ public class GameScreen extends Screen {
         }
         this.finalBoss.update();
 
-        bossAttacks.addAll(this.finalBoss.shootBossAttack());
+        bossAttacks.addAll(this.finalBoss.processAttacks());
+
+        if (this.finalBoss.shouldClearAttacks()) {
+            bossAttacks.clear();
+        }
 
         manageBossAttacks();
     }
-        /* 
-        if (this.finalBoss.getHealPoint() > this.finalBoss.getMaxHp() * FinalBoss.PHASE_2_HP_THRESHOLD) {
-            if (this.finalBoss.getDifficulty() == 1) {
-                bossBullets.addAll(this.finalBoss.shoot1());
-                bossBullets.addAll(this.finalBoss.shoot2());
-            } else {
-                bossBullets.addAll(this.finalBoss.shoot3());
-            }
-        } else if (this.finalBoss.getHealPoint() > this.finalBoss.getMaxHp() * FinalBoss.PHASE_3_HP_THRESHOLD) {
-            if (this.finalBoss.getDifficulty() != 1 && !is_cleared) {
-                bossBullets.clear();
-                is_cleared = true;
-            } else {
-                bossBullets.addAll(this.finalBoss.shoot1());
-                bossBullets.addAll(this.finalBoss.shoot2());
-                bossLasers.addAll(this.finalBoss.laserShoot());
-            }
-        } else {
-            if (this.finalBoss.getDifficulty() != 1) {
-                bossBullets.addAll(this.finalBoss.shoot4());
-            }
-            bossBullets.addAll(this.finalBoss.shoot2());
-        }
-
-        Set<BossBullet> bulletsToRemove = new HashSet<>();
-
-        for (BossBullet bossBullet : bossBullets) {
-            bossBullet.update();
-            if (bossBullet.isOffScreen(width, height)) {
-                bulletsToRemove.add(bossBullet);
-            }
-            else if (this.lives > 0 && this.checkCollision(bossBullet, this.ship) && !GameState.isInvincible()) {
-                if (!this.ship.isDestroyed()) {
-                    this.ship.destroy();
-                    this.lives--;
-                    this.logger.info("Hit on player ship, " + this.lives + " lives remaining.");
-                }
-                bulletsToRemove.add(bossBullet);
-            }
-        }
-
-        Set<BossLaser> lasersToRemove = new HashSet<>();
-
-        for (BossLaser bossLaser : bossLasers) {
-            bossLaser.update();
-            if (bossLaser.isRemoved()) {
-                lasersToRemove.add(bossLaser);
-            }
-            else if (this.lives > 0 && this.checkCollision(bossLaser, this.ship) && !GameState.isInvincible()
-                    && takeLaserDamageCooldown.checkFinished()) {
-
-                takeLaserDamageCooldown.reset();
-
-                if (!this.ship.isDestroyed()) {
-                    this.ship.destroy();
-                    this.lives--;
-                    this.logger.info("Hit on player ship, " + this.lives + " lives remaining.");
-                }
-            }
-        }
-        bossBullets.removeAll(bulletsToRemove);
-        bossLasers.removeAll(lasersToRemove);
-        */
 
     /**
      * Manages the boss attacks, updating their positions and checking for collisions with the player ship.
@@ -1056,7 +982,10 @@ public class GameScreen extends Screen {
                     this.lives--;
                     this.logger.info("Hit on player ship, " + this.lives + " lives remaining.");
                 }
-                attacksToRemove.add(bossAttack);
+
+                if (bossAttack instanceof BossBullet) {
+                    attacksToRemove.add(bossAttack);
+                }
             }
         }
         bossAttacks.removeAll(attacksToRemove);
